Create a function
Create a function which take two arguments. One for country name and another for column name from survey.csv. First, this function will filter data according to country name and store that data into country_data variable. After that it will group by this data according to second argument and take count of it. Last it will plot the graph